Small reliability fixes

+++ app/src/main/java/io/xpipe/app/core/check/AppUserDirectoryCheck.java
@@ -15,14 +15,15 @@ public class AppUserDirectoryCheck {
             Files.createDirectories(dataDirectory);
             var testDirectory = dataDirectory.resolve("permissions_check");
             Files.createDirectories(testDirectory);
+            if (!Files.exists(testDirectory)) {
+                throw new IOException("Directory creation in user home directory failed silently");
+            }
             Files.delete(testDirectory);
             // if (true) throw new IOException();
         } catch (IOException e) {
-            ErrorEvent.fromThrowable(
-                            new IOException(
-                                    "Unable to access directory " + dataDirectory
-                                            + ". Please make sure that you have the appropriate permissions and no Antivirus program is blocking the access. "
-                                            + "In case you use cloud storage, verify that your cloud storage is working and you are logged in."))
+            ErrorEvent.fromThrowable("Unable to access directory " + dataDirectory
+                            + ". Please make sure that you have the appropriate permissions and no Antivirus program is blocking the access. "
+                            + "In case you use cloud storage, verify that your cloud storage is working and you are logged in.", e)
                     .term()
                     .expected()
                     .handle();
